Improved overlay

+++ SR-Overlay/SDK/GameSDK.hpp
@@ -44,17 +44,17 @@ namespace SDK {
 	typedef struct usercmd_s
 	{
 		short	lerp_msec;      // Interpolation time on client
-		byte	msec;           // Duration in ms of command
+		int	msec;           // Duration in ms of command
 		vec3_t	viewangles;     // Command view angles.
 
 		// intended velocities
 		float	forwardmove;    // Forward velocity.
 		float	sidemove;       // Sideways velocity.
 		float	upmove;         // Upward velocity.
-		byte	lightlevel;     // Light level at spot where we are standing.
+		int	lightlevel;     // Light level at spot where we are standing.
 		unsigned short  buttons;  // Attack buttons
-		byte    impulse;          // Impulse command issued.
-		byte	weaponselect;	// Current weapon id
+		int    impulse;          // Impulse command issued.
+		int	weaponselect;	// Current weapon id
 
 		// Experimental player impact stuff.
 		int		impact_index;
@@ -82,8 +82,8 @@ namespace SDK {
 		// Complex collision detection.
 		float			frame;
 		int				sequence;
-		byte			controller[4];
-		byte			blending[2];
+		int			controller[4];
+		int			blending[2];
 
 		int				movetype;
 		int				takedamage;
